[augmentation] Check for ints, not floats, in padding output size tuples

+++ FastSurferCNN/data_loader/augmentation.py
@@ -48,7 +48,7 @@ class ZeroPad2DTest(object):
         :param output_size:
         :param pos: position to put the input
         """
-        if isinstance(output_size, float):
+        if isinstance(output_size, int):
             output_size = (output_size, ) * 2
         self.output_size = output_size
         self.pos = pos
@@ -107,7 +107,7 @@ class ZeroPad2D(object):
         :param output_size:
         :param pos: position to put the input
         """
-        if isinstance(output_size, float):
+        if isinstance(output_size, int):
             output_size = (output_size, ) * 2
         self.output_size = output_size
         self.pos = pos
